requirements.txt
discord.py
spawningtool
python-dotenv
